kbuild: Allow .dtso format for overlay source files
Since the overlays dtb files are now named as .dtbo, there is a lot of
interest in similarly naming the overlay source dts files as .dtso.

This patch makes the necessary changes to allow .dtso format for overlay
source files.

Note that we still support generating .dtbo files from .dts files. This
is required for the source files present in drivers/of/unittest-data/,
because they can't be renamed to .dtso as they are used for some runtime
testing as well.

Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>
Tested-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Link: https://lore.kernel.org/r/170e086a5fa076869e7b37de8eea850fa7c39118.1615354376.git.viresh.kumar@linaro.org
[geert] Dropped first hunk, obsoleted by commit 64bfc99429a5c961 ("kbuild: remove unneeded -O option to dtc")]
[geert] Rebase on top of commit bcf0c66428336738 ("kbuild: refactor fdtoverlay rule") in v5.13-rc1.
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
(cherry picked from commit 2dd668cd49762ccdac55ec7549d8ba80ffb71d70)