gpio: Support GPOs on ROHM BD2657 PMIC
The ROHM BD2657 has two optional GPIOs. Pins can also be reserved for
other purposes. Additionally, the input functionality is mostly
undocumented so exopose pins only as GPOs.

Set the GPIO0 under HW control if it is excluded by reserved-ranges
property. (This is something the boot or MFD should probably do). I may
add own propert for this and change this.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>