power: supply: simple-gauge: Fix clamped SOC
The clamped_soc was not updated before battry was charging. Clamped
SOC is increased only if charger state is CHARGING as else the SOC
increment is assumed to be caused by ADC offset. However the clamped
SOC was initialized to zero so and stayed at zero untill the first
charging.

Initialize clamped_soc to invalid value and update it to computed
SOC at first cycle (even if battery was not charged).

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>