power: supply: bd71827: Fix Coulomb Counter to uAh conversion
The multiplication in coulomb counter to uAh conversion may overflow
with large CC values. Fix this by decreasing the multiplier and
increasing the divider at the next calculation. Some accuracy may
be lost but thats far better than risking to overflow.

Additionally the conversion macro was intended to operate only on
the high 16 bits of the CC - but it was errorneously used with
full 32bit CC. Do separate clearly named macros for 16 and 32 bit
conversions to avoid such errors.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>