pmic-tests: Add initial version of scripts
Add the collection of scripts for testing the Linux drivers for
some of the ROHM PMICs.

The tests expect an environment consisting of a Linux PC, two
BeagleBone blacks, (ethernet switch), remote-controllable
power-outlet and various PMIC break-out boards.

Inaccurate picture to illustrate the potential setup.

		 ------- ------
		 |PMIC | |PMIC|
		 ------- ------
		       | |i2c & power
		       | |
		     __|_|__
--+---	______	     | BBB1| eth
| PC |  |IP   |------|	   |------+
|    |--|power|	DC   =======	  |
|    |  |     |------| BBB2| eth  |
|    |  -------	     |     |---+  |
|    |		     --+-+-- __|__|__
|    +-----------------|-|---|switch |
|    |	ethernet       | |   ---------
--+---		       | |
		       | |i2c & power & ADC
		       | |
		 ------- ------
		 |PMIC | |PMIC|
		 ------- ------

The PC hosts these tests and can host some automated test system
which launches the test set when a new Linux release is created.

These test scripts should fetch the given Linux release from
upstream repository and cross-compile the Linux kernel + test
modules. After compilation the scripts trigger reset on connected
beaglebone black - which boots up. The BBB uBoot should be
configured to load the device-tree and linux image from PC.
Additionally the PC should host NFS which the BBB uses as rootfs.

After BeagleBoneBlack has booted the scripts add device-tree overlay
for PMIC to be tested + modprobe required test module and finally
execute the test cases on beaglebone black. The test cases do
access the connected PMIC via I2C and ensure voltage settings and
enable/disable requests issued via Linux regulator framework and
the PMIC drivers are reflected in PMIC's registers. Additionally
some voltage outputs from some of the PMIC break-out boards are
wired to the ADC inputs on BBB. For these outputs also the actual
voltage change is verified. This measurement testing is not
prioritized and quite big tolerances are accepted as this is not
hardware testing and the test wiring may not be of best quality.
...I am a software guy after all ;)

This commit only adds the test scripts and some PMIC dts / test
module files. The uBoot env for tftpboot and NFS rootfs are still
stored in another repository.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>