jenkins-scripts: further setup.sh work
The setup.sh should initialize git repositories and create folders.
It should also prepare the environment variables which are used by
the test scripts so that they find the PC specific paths. Here we
continue setup.sh development and take the environment variables in
use at the test scripts.

Changes here include (but are not limited to)

- The test kernel modules, test scripts and test DT blobs should be available
  together with the tested kernel sources. Checkout them in their own dir.

- CHeckout the compilers etc. from tetrao-urogallus in case they are
  not found elsewhere.

- Do few generic functions like one to ask user what to do, and one to
  set-up git remote.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>

---
This is not-yet-working test commit which will be rebased.
We do intermediate commits here because the script works by fetching
configs from this remote repo - and we need the configs updated
while testing/developing the scripts.