jenkins-scripts: setup.dh: Don't overwrite configs
The setup.sh did ask if user wanted to maintain existing config files
but still unconditionally appended the new RUNENV configs. This is not
maintaining old configs and leads to multiple RUNENV configs when old
configs are asked to be maintained.

Dont add RUNENV configs if old config is asked to be maintained.

Note, if new version adds new RUNENV configs they are no longer appended
to existing configs. This means that if the backwards compatibility is
not maintained (and I don't think we should right now invest on that at
all) then the user _should_ back-up old configs - create new ones, and
manually merge the configs he wants to keep from old setup.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>