jenkins-tests: speed-up the setup.sh
When setup.sh is ran to just update existing environment (mostly when
developing setup script or adding new env variables) the most of the
time is spent by checking if a remote repository is alrady added. The
existin check does this by querying the remote - which can be slow for
large remotes like the kernel.

Let's speed-up this by just doing a quick remote check by checking the
get-url returns 0. It only checks if remote with given name exists.
moreoverm it does not say a thing about remote validity -  but it is
_much_ faster than the remote show.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>