jenkins-tests: Fix the cross-compiler setup
The script which sets up the compiler environment (setcc) was hosted in
the tetrao-urogallus repository together with the cross-compiler. Change
this so that the RUNCFG_COMPILER_ENV_PATH environment variable is
pointing to the file which sets the compiler environment - and utilize
this env variable from the build files.

Additionally fix the log directory - (sigh. another commit that
should've been two separate changes)

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>