jenkins-scripts: Out of tree module build script
The test modules for PMICs are stored as out-of-tre modules. This helps
us to avoid Kconfig and Makefile conflicts which we would probably have
if we hosted the modules in-tree and built them using the in-tree
Kconfig.

Add build script which sets the cross-compiler and kernel source
directory prior calling the Make.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>