pmic-tests: Add overlay-adding mechnism
We want to add DT blobs representing ICs we test to running kernel
in order to instantiate the drivers for PMIC which we are testing.

Add overlay adding code which supports inserting compiled DTB blob
via sysfs.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>

---

Note: This code is not actively maintained. As one indication about
that, removing a DT overlay blob is not working. I don't currently
need that functionality - so I haven't bothered to fix it.

If you wish to have a maitained version of overlay-adding from
user-space, then I suggest you go get the one from Renesas tree.
That code is maintained by Geert and bunch of people seem to be
contributing to that. I hope I some day find the time to get rid of this
version and convert the ROHM tests to fetch the patches from Renesas
tree...