power: supply: Don't assume 32bit int
The battery-info assumes the 32-bit DT properties can be directly read
into a table of ints. This should probably work for pretty much every arch
running linux we have out there - but it is not mandated. So let's not
assume.

By the way, I was once in a training for trainers - where the trainer
explained me where assume comes from. Did you know that when you assume,
you make an ASS out of U and ME?

Don't do that.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>