mfd: BD96801: Add registers which cant always be written to volatile ranges
A few of the configurations like regulator enable/disable state and
protection conifgurations can be only written when PMIC is in STBY
state. If writing is done when PMIC is in ACTIVE (or any other state but
STBY), the write will be seemingly successfull but the value contained
in register does not change. Thus writing any of the "STBY-only"
registers in any other state will make cached register value and the
real value contained in HW to be different.

With BD96801 this may play a role as it seeems we should prepare to
setups where processor is powered by some output which is enabled when
PMIC is in STBY. Eg, SW may be running and should be able to change PMIC
configs also when the PMIC is in STBY. So we need to also implement the
access to the registers which can be only written when PMIC is in STBY.

Prevent the cache mismatch problems by marking volatile the registers
which can be changed only at STBY mode. Then we at least do read
correect values if we write registers at wrong state (and we can
implement a read-back to confirm registers were updated if needed).

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>