jenkins-test: bd718x7: disable unreliable voltage measurement tests
The voltage measurement tests are not relaible. Occasionally the voltage
readings from BBB ADC are badly off. This may be caused by bad cabling.
Or ramp-delay/measurement time mismatch. Or aomw electrical phenomena
which we could work-around by capacitors / delays / measurement averaging
or by aligning the stardust generator with the proper warp-zone.

None of this really matters. We don't try to do hardware testing here.
We do device driver software testsing trying to get a warning if
upstream kernel change breaks our driver. All the (regulator) software
does is writing new values to the PMIC registers when regulator framework
is used to change the output. So what we care is that register values
written to PMIC match the configurations done by rhe test code (via
regulator framework and our driver). And the test does verify the register
values using I2Ctools. The physical voltage measurement is something
that should not be needed at regression testing - we do that once when
we ensure we understood the HW colleagues correctly - but we do not need
to do this when the Linux kernel is changed as long as the change is not
done because our expectations how register values impact voltages are
not changed.

After this long rambling just remove the AIN measurements form the
tests.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>