Revert "regulator: bd96801C: Start work for allowing config phase to turn PMIC to STBY"
This reverts commit 124198cddaddf3d96ee0825785c098e47ee412d8.

The generic regulator driver setting PMIC to STBY was a doomed idea. If
a runtime configuration needs to be done and PMIC needs to be set to STBY
for it - then it is a job for system specific consumer driver, not a job
of regulator driver which aims to be a generic one.

If initial configs need to be done at STBY - then setting PMIC to STBY
at early stage is a task for bootloader. Activating PMIC after configurations
are done _could_ be performed by PMIC driver - we may wish to add a DT
flag for this. The PMIC driver knows when all of the regulators are
registered to the core (and core has completed initial configs).

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>