watchdog: bd96801: Always initialize hw_heartbeat
The BD96801 watchdog initializes the hw_heartbeat when it configures the
desired hardware time-out. If watchdog is enabled at the probe time, the
driver assumes that boot has been already configuring the time-outs and
other hardware parameters. The driver then skips doing the
configurations.

In this case the hw_hearbeat is not configured. This is likely to lead
to reboot should the timeout configured by boot be tight.

Read the configured timeouts when the watchdog is enabled at boot and
set the hw_heartbeat based on the values set to the timeout registers.

Fixes: b209c2845956 ("watchdog: ROHM BD96801 PMIC WDG driver")
Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>