watchdog: bd96801: drop DT specific APIs
the direct use of of_*() APIs is discouraged as it means the frmware
back-end must be device tree. fwnode_*() APIs should work on other systems
like ACPI as well. This should help if the driver is ever used on anything
else but device-tree backed systems.

Use the fwnode_*() APIs instead of of_*().

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>