web: bd96801: Clarify that we have initial driver
The statement that the Linux driver is under development is slightly
misleading. It may suggest that customers should not take the initial
releases but to wait for a more mature one. This is not the case as
even though there is some development ongoing, there is no accurate plan
as to when we generate new release.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>