watchdog: bd96801: Support ROHM BD96802 watchdog
The ROHM BD96802 companion PMIC has similar wtachdog block as BD96801
main PMIC does. No differencies were spotted from specification but they
are fundamentally different as the BD96802 is mainly intended to be used
as a 'companion PMIC' which startup/shutdown can be controlled by the
BD96801 although it is possible to also use BD96802 as a stand alone
PMIC. This may be resulting some differencies which are not yet visible
in the spec version I have at my hands. So, it is better to have own IDs
for those PMICs to be able to easily differentiate between them in case
this is needed.

Add support for ROHM BD96802 PMIC's watchdog block.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>