mfd: bd96801: Support ROHM BD96811
The ROHM BD96811 is an automotive grade PMIC that can be heavily
customized by OTP. One of the customizations is the presence of an IRQ
pin. An DOUT pin can be configured to serve as an IRQ pin by OTP.

I have at my hands does not have the IRQ pin so the IRQ support for
BD96811 is not implemented.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>