watchdog: bd96801: Support ICs with different WDG base address
The ROHM BD96811 PMIC has a watchdog register interface which is almost
exactly the same as the interfaces on ROHM BD96801 and BD96802. Only
difference is the start address of the watchdog logic block. It is
straightforward to support ICs like this by allowing just specifying the
start adress of the watchdog block and computintg the register addresses
as offsets from the start.

Support giving per IC watchdog register block start address.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>