regulator: Support ROHM BD96811
ROHM BD96811 is an automotive grade PMIC that contains 5 voltage outputs.
The type of the outputs can be configured by OTP. Vout1 is always a
BUCK, Vout 2 can be a BUCK or a BOOST, Vout 3 and 4 can be BUCKs or LDOs
and Vout 5 is always an LDO.

The configuration can't be read from PMIC by SW so driver relies on
correct type being given from device-tree.

The BD96811 does also support over/under-voltage and over-current
protection/detection. The IRQ to inform errors is optional as an DOUT
pin can be configured to serve as an IRQ pin by OTP. The current variant
I have at my hands does not have the IRQ pin so notification sending is
not implemented. The OVP/UVP/OVD/UVD limit configuration is supported
but the HW has a few caveats:

1. OVD/UVD limit 'base value' can be configured by OTP but is not
   readable by SW. Hence, if limits are to be configured, the base limit
   that is set by OTP must be given to driver via device-tree.

2. OVD/UVD can be changed to be 'fatal'. Eg, there is an option to make
   OVD/UVD forcibly shut down the power outputs. This effectively
   changes the OVD/UVD to OVP/UVP. Driver is utilizing this setting if
   OVP/UVP limit is being changed to a value which can be supported by
   OVD/UVD. The downside is that this 'fatality' configuration is
   global. When 'fatality' is enabled, all detection features (both
   over and under voltage as well as thermal warning) will change to
   fatal. As a small safety feature the driver will warn and error if
   both warning/error and protection limits are attempted to be used.

3. The OVD and UVD settings are combined into single register field so
   that setting the UVD limit will also impact the OVD limit and the
   other way around. Driver will again error out if both OVD and UVD
   limits are tried to be configured in a way that the resulting
   register value is not exactly same.

Support ROHM BD96811 regulators.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>