dt-bindings: mfd: bd96811: Add intb IRQ
One of the ROHM BD96811 PMIC DOUT pins can be configured to work as an
INTB IRQ pin by OTP. Complete the device-tree binding.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
This is really not the right thing to do. The BD96811 does not
differentiate the INTB and ERRB IRQs but just provides a single IRQ.
Thus, the interrupt-name usage is not really correct. This is added here
to simplify the driver - the BD968[01/02/11] share the IRQ code so it is
simplest though incorrect to require use the same names.

TODO: Fix this by fixing the IRQ getting code in mfd driver prior
upstreaming.