mfd: bd96811: IRQ support
ROHM BD96811 has an OTP option to use one of the DOUT pins as an IRQ.
The BE96811 differs from BD96801 and BD96802 as it only has one IRQ pin,
not the INTB and ERRB separated pins as BD96801 and BD96802.

Add support to BD96811 IRQs in MFD core driver.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
NOTE: Completely untested code.
NOTE: Because the BD96811 has only 1 IRQ pin, it could have simple and
clean approach without the explicit IRQ mapping in MFD driver. This
should be fixed if driver is upstreamed.