pinctrl: Support pinmux for ROHM BD96811
ROHM BD96811 PMIC has DOUT and DIN pins for which the functionality can
be muxed. Specifically the DOUT4 and DIN4 can be either operating as ADC
inputs or as specific DIN/DOUT purpose defined by OTP.

The primary use for these ADCs is monitoring voltages of discrete
voltage sources and thus the muxing the ADC to the pins is tied to ADC
safety threshold voltage configuration. When safety limit is zero, these
pins are muxed to DIN/DOUT functionality defined by OTP. When limit is
non zero the pinmux configures ADC.

Some of the DIN/DOUT pins can be configured as GPOs by OTP. The OTP
setting can not be read from the device so information about the GPOs
should come from device-tree. This version of the pinconf driver does
not support the GPO configuration.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>