mfd: bd96811: Prepare for ADC support
Prepare for supporting the BD96811 ADC by adding the ADC pin muxing
registers to header. These registers are needed by both pinctrl and adc.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
This commit is independent because - for the ROHM downstream repoistory
it is an addition to already released 'bd96811 alpha release'. As an
independent commit this can be applied on top of the alpha release
without conflicts. So, we don't squash this commit into original commit
introducing the BD96811 MFD core (which it would really belong).

On the other hand, we could squash this to the commit introducing either
the pnctrl or the ADC drivers. However, for potential upstream
submission we want to go with clear separation of changes to MFD and
changes to sub-drivers. Hence, for upstream we want to squash this in
introduction of the BD96811 MFD core.

So, here we have this as independent commit which allows it to both be
applied on existing woenstream alpha release and also allows easy
squashing to MFD core if I am ever allowed to upstream the scalable PMIC
drivers...