mfd: bd96811: Support pinctrl and ADC
The BD96811 has a DOUT4 and DIN4 pins which can be used wither for the
DOUT/DIN functionality defined via OTP or as an ADC input. A pinctrl
driver which can enable ADC using pinmux and an IIO ADC driver which can
return the measured voltages exist in respective subsystems.

Enable the MFD cell for pinctrl and ADC drivers.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>