regulator: bd71828: Prepare for supporting BD72720
The BD72720 will have plenty of commonalities with the BD71828. It makes
sense to support both PMICs with the same regulator driver. Current
driver does directly refer to the BD71828 regulator desc table. Change
the logic to use pointer to table so that new variants can just set the
pointer to correct desc tables in the probe.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
100% untested