power: simple-gauge: drop static from SIMPLE_GAUGE_PROPS
The array SIMPLE_GAUGE_PROPS is defined as a static in the header being
included by the drivers using the simple-gauge. The SIMPLE_GAUGE_PROPS
is intended to be unmodified and it is not justified for every driver
which includes the simple-gauge.h to have own copy of the array.
Furthermore, all of the drivers which include the header aren't using
the SIMPLE_GAUGE_PROPS array, which will cause a compiler warning about
unused array when the array is static.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
Is there actually any use for this array outside the simple-gauge.c?
Should it be taken out of the header alltogether? I can see it serving
some 'informative purpose' as it lists the power-supply entities being
sent out to userspace by the simple-gauge.