gpio: Support ROHM BD72720 gpios
The ROHM BD72720 has 5 pins which may be configured as GPIOs. The
configuration is done to OTP at manufacturing, and it can't be read at
runtime. The device-tree is required to tell the software which of the
pins are used as GPIOs.

Two first pins can be set to be either input or output by OTP. Direction
can't be changed by software. Three of the pins can be set as outputs
only. All of the pins support generating interrupts.

Support the Input/Output state getting/setting and the output mode
configuration (open-drain/push-pull).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
100% untested.
No interrupt support yet