mfd: bd72720: Support being irq-controller
The BD72720 has two pins which can be used as IRQ input pins. Eg,
the BD72720 can be an IRQ parent for other devices. The interrupt
type (fallins, rising and boths edges as well as high and low levels).

Add support for using BD72720 as an interrupt controller and configuring
the interrupt types.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
100% untested