led: bd71828: Support ROHM BD72720
ROHM BD72720 and BD71828 have identical LED control logic. Extend the
BD71828 driver to support also the BD72720.

NOTE: The LED control for these ICs is not trivial. The LEDs can be
either 'forced' (Eg, software can lit the LED), or in a 'automatic' mode
where the hardware state is automatically indicated by the LEDs.

This driver provides control by 'forcing' the LED(s) 'ON'.

There is a twist. The state of the both LEDs is same. Eg, they are both
either in 'forced', or in 'automatic' state. Furthermore, the state
control is bound to LED enable state. The state of both LEDs is changed to
forced, if one of the LEDs is turned ON by the register.

Implication is that turning one LED ON will disable the 'automatic' mode
for the other LED as well. It means a previously 'hardware state
controlled' other LED will now be forced OFF. It also means there is no
way to force both LEDs OFF - because setting them both OFF will enable
the 'automatic' mode - and LED status will then reflect the hardware
state.

This driver allows forcing LEDs on via the Linux LED interfaces

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% untested