power: supply: bd71828: Fix Rsense resistor value
The sense resistors on the ROHM PMICs supported by the bd71827-power
driver are typically between 10 and 50 milli volts. Due to a series of
misunderstandings, the code / device-tree have assumed the resistors are
of size 10 to 50 mega ohms.

Drop reading of the 'rohm,charger-sense-resistor-ohms' which is used to
give 1 000 000 000 times too large resistor values, and which can't be
used to give correct (less than one ohm) values. Add support for reading
'rohm,charger-sense-resistor-milli-ohms' instead.

Also, fix the computations accordingly.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
100 % untested