dt-bindings: mfd: bd71828: Fix sense resistor values
The sense resistor used for coulomb counting is really of size of tens
of milli Ohms. It has been accidentally tought of being aroung tens of
mega Ohms.

Fix this by deprecating the old "rohm,charger-sense-resistor-ohms"
property which does not suit representing milli ohm scale resistors and
introduce new "rohm,charger-sense-resistor-milli-ohms" for this.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Fixes: 1af5332fcf7c ("dt-bindings: mfd: Document ROHM BD71828 bindings")