dt-bindings: Add trickle-charge upper limit
Some of the chargers for lithium-ion batteries use a trickle-charging as
a first charging phase for very empty batteries, to "wake-up" the battery.
Trickle-charging is a low current constant current phase. After the
voltage of the very empty battery has reached an upper limit for
trickle charging, the pre-charge phase is started with higher current.

Allow defining the upper limit for trickle charging voltage, after which
the charging is changed to the pre-charging.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
100% untested