mfd: bd71828: Definition for fast charge term current register
The PMIC charger blocks supported by the BD71827-charger driver do all
support configuring various aspects of charging phases. One of the
specifics is the currennt value when constant voltage charging phase is
ending. Eg, typically the "main charging phase" (called fast charging in
the various ROHM PMIC data-sheets) is divided in two phases. First being
constant current phase, when battery is charged keeping current
constant, and allowing the voltage to rise. When certain voltage level
is reached, the charging is changed to a constant voltage phase, where
the charging voltage is kept constant untill the current flowing in the
battery has dropped down to a certain level (and battery is usually then
considered full). This termination current can be specified.

Add the definition of the BD71828 PMIC's register which handles
configuring of the termination current.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>