doc: Add a pdf doc for bd72720 stuff
Yeah. This does not really belong to the sources, but I hope this gives
a hand for anyone who actually tries to understand what the BD72720
drivers do. It's not something that fits the ipstream code though, so we
should find a place where to put a polished version of it (rohm github
pages perhaps?) - and we could then link to this doc from the code
files.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>