web: bd96801: Drop downstream simple-driver
The downstream drivers are not maintained. They have unfixed bugs and
won't be ported to use new safer mechanisms. They were never meant to be
used as such - they were provided as a reference designs only.

Currently we have a bd96801 driver upstream. The upstream driver is
better maintained, and provides all the same functionality the
simple-driver did. All users should prefer it, and we should no longer
promote the suboptimal version.

Drop the simple-driver from the BD96801 info.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>