regulator: bd71891: Prevent range change if enabled
The data-sheet states that the voltage ranges must not be changed when
voltage output is enabled. Return -EBUSY if this is attempted.

NOTE: There is no locking in the driver, so enable check is racy. The
serialization of the voltage setting / enabling calls must be done by
the calling code (or by the system).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% Untested