pmic: bd71891: Add chipinfo command
Add a debugging command for printing PMIC state information. The command
prints:
 - Stored PMIC boot diagnostics
 - Current internal state machine state.

This should be helpful both for debugging, and also for serving as a
code example about reading the reset / boot state registers.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% Untested. Also, it may be the RESETSRC2_FLG and RESETSRC3_FLG need
only to be read if BIT(6) is set in the RESETSRC1_FLG register. This,
however, is a minimal optimization.