pmic: bd71891: Add HDMI HPD
The IDLE state can be controlled by HDMI hot-plug detection. Add check
for this in idle setting command because the IDLE mode changing does not
work if this is controlled by the hardware.

Also, it can be helpful for debugging to be able to check the state of
the HDMI HPD IDLE control. Additionally, having a control for this
servers as an example for customers who may want to control it.

Add commands to query the status and enable/disable HDMI hot-plug's
IDLE control.

Add also commands for HDMI pin configuration for all the same reasons.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% Untested.