power: bd71891: Unlock I2C at probe
Registers of the BD71892 are locked (writing is prevented). A magic
sequence needs to be written to a lock register, in order to allow
updating the values.

The regulator control (as well as other controls) require the registers
to be written.

Unlock the registers by using the magic writes to the lock register. Do
this in the pmic driver, before binding the sub drivers (which could fail
because register writes from sub-drivers wouldn't work as expected).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>