pmic: bd71891: Add ADC accum voltage source mux control
The BD71891 accumulator has two multiplexers so, that when the one
closest to the accumulator is set to measure voltage, then the second
can be used to switch between different voltage sources. In the case of
the BD71891, possible sources are GATE_DRV and VSYS.

Users may need to be able to switch between the sources, so add a
command which can be used as a code example to do the switching.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% untested