pmic: bd71891: Add ADC measurement
The BD71891 adds users a way to set the ADC to take measurements at the
background. This can be useful when longer term average value is needed,
or when things need to be monitored.

Add a command for taking N samples of given type of measurements so
users can study and experiment this.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% untested