pmic: bd71891: Support USB characterization
Add two U-boot commands for USB characterization.

One for reading the results (and optionally starting the measurement):
'bd71891 usb_char [start]'

and the other for getting / setting the USB characterization
configuration:
'bd71891 usb_char_cfg [config value]'

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% untested.

As of now this only returns the raw results as read from the registers.
Converting the register values to meaningful units is not (yet) done.