power: bd71892: Drop BD71892 GPIO specific definitions
A few of the register bit positions/values used for the GPIO handling are
same for the BD71891 and BD71892. Common defines were added for these.
Since the BD71892 specific defines are now unused, we can clean up the
header a bit and drop the unused defines.

Drop unused defines.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
100% Untested