dts: Test PMIC node fixups
Fix_1:

The u-Boot version used here seem to have some problems handling
multiple regulators with identical names. The u-Boot regulator class
commands use the name to distinguish devices.

The BD71891 and the BD71892 both have regulators called buck1, buck2,
and so on. The u-Boot regulator commands fail to list the duplicate
regulators.

Use capital names for one PMIC, and non capital ones for the another
PMIC to distinguish the similarly named regulators on the PMICs.

Fix_2:
The u-Boot regulator shell commands do not allow changing voltages
without -f (force) flag if there are no voltage constrains set in the
device-tree.

Add dummy voltage constraints by setting minimum and maximum same as the
PMIC's minimum and maximum.

NOTE: This is only for testing. In real systems the constrains should be
chosen based on what the hardware which is connected to the PMIC can
handle. Eg, these constrains are intended to protect the hardware, not
to follow what the PMIC can do.

NOTE2:
The BD71892 DS1 had some problems with the BUCK2. Hence the BUCK2
voltages are shifted upwards by 20 mV for low range, and 40 mV for high
range. The BUCK2 constrains here follow the DS1 min and max.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>