regulator: bd71892: Fix the regulator probing.
The regulator device name is used in the regulator probe to identify the
currently probed regulator. The current test device-tree results
regulator names for the BD71892 to be lowercase. The driver however, has
the regulator names in uppercase, and fails matching the regulator data
to the device.

We don't want to change the device tree to result capital names for
BD71892 to keep the binding prefix similar for BD71891 and BD71892, while
still allowing the u-Boot to still distinguish the BD71891 and BD71892
regulators.

Make matching the device and regulator data working by using lower-case
names for matching in the driver.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>