gpio: bd718xx_gpio: Fix direction setting
The BD71891 and BD71892 GPIO mode setting is implemented using the PMIC
class register accessors. This requires the GPIO driver to provide a
pointer to the PMIC device as an argument to register read/write
functions. BD7189x GPIO devices are registered as children of the PMIC
device. Thus the pointer to the PMIC device can be obtained by obtaining
the pointer to the GPIO device's parent device.

The mode setting code path did obtain the parent device pointer in a
higher level function, passing it to lower level functions. These lower
level functions expect pointer to the GPIO device, not the parent. This
caused the parent of the PMIC device to be passed as an argument to the
pmic-class register accessor's. This can cause severe problems as the
pmic class expects the device to have proper private data.

Fix the higher layer code to pass the GPIO device's address as expected,
and obtain the parent (PMIC) device's pointer only in the lower layer,
eventually passing the correct device pointers to register accessor
functions.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>